pauli circtuit into dedicated module
